+++ lib/project_types/script/cli.rb
@@ -50,6 +50,7 @@ module Script
       autoload :ExtensionPoint, Project.project_filepath("layers/domain/extension_point")
       autoload :ScriptConfig, Project.project_filepath("layers/domain/script_config")
       autoload :ScriptProject, Project.project_filepath("layers/domain/script_project")
+      autoload :AppBridge, Project.project_filepath("layers/domain/app_bridge")
     end
 
     module Infrastructure