+++ lib/project_types/script/layers/application/push_script.rb
@@ -55,6 +55,7 @@ module Script
                   force: force,
                   metadata: package.metadata,
                   script_config: package.script_config,
+                  app_bridge: script_project.app_bridge,
                   module_upload_url: module_upload_url,
                   library: package.library,
                   input_query: script_project.input_query,